html/browsers/browsing-the-web/navigating-across-documents/009-onclick-form-action-javascript-url-document-write-script-then-navigate-href.html

<!doctype html>
<link rel="help" href="https://html.spec.whatwg.org/multipage/browsing-the-web.html#navigating-across-documents">
<link rel="help" href="https://html.spec.whatwg.org/multipage/browsing-the-web.html#javascript-protocol">
<link rel="help" href="https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#concept-form-submit">
<link rel="help" href="https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#submit-get-action">
<link rel="help" href="https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#plan-to-navigate">
<link rel="help" href="https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#planned-navigation">
<title>Link with onclick form submit to javascript: url with document.write and href navigation</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<iframe src="017-onclick-form-action-javascript-url-document-write-script-then-navigate-href-intermediary.html"></iframe>
<script>
var test = async_test();
window.onmessage = test.step_func_done(function(e) {
  assert_equals(e.data, "href", "'javascript: url' code is expected to run asynchronously and in the target nested browsing context");
});
</script>


<!--
In an iframe...

1. This test creates an iframe, a form with
    `action="javascript: ..."` (which document writes a script and returns a strings
    containing a script. Both scripts attempt to postMessage back to the intermediary.html)
    and an anchor with both an `onclick` and `href`.
2. The test "clicks" the anchor, triggering the `onclick` handler operation,
    which will attempt to submit the form, which targets the iframe.
3. The form action "javascript: url" navigation is queued in a task which contains the
    steps to evaluate the code contents.
4. Navigating the iframe to the anchor's `href` must cancel the iframe
    "javascript: url" navigation.
5. If the semantics are implemented correctly, a single message will be
    received from href.html
    (There will be no message from the "javascript: url" code, because it
    is aborted by the navigation to "href.html")
-->
